+++ gradle-plugin/src/proguard/gradle/ProGuardTask.java
@@ -2,7 +2,7 @@
  * ProGuard -- shrinking, optimization, obfuscation, and preverification
  *             of Java bytecode.
  *
- * Copyright (c) 2002-2020 Guardsquare NV
+ * Copyright (c) 2002-2021 Guardsquare NV
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by the Free
@@ -24,12 +24,17 @@ import groovy.lang.Closure;
 import org.gradle.api.DefaultTask;
 import org.gradle.api.file.*;
 import org.gradle.api.logging.*;
+import org.gradle.api.model.*;
+import org.gradle.api.provider.*;
 import org.gradle.api.tasks.*;
+
 import proguard.*;
 import proguard.classfile.*;
 import proguard.classfile.util.ClassUtil;
+import proguard.gradle.configuration.ProGuardTaskConfiguration;
 import proguard.util.ListUtil;
 
+import javax.inject.Inject;
 import java.io.*;
 import java.net.*;
 import java.util.*;
@@ -51,10 +56,10 @@ public abstract class ProGuardTask extends DefaultTask
     private final List          inJarCounts        = new ArrayList();
     private final List          libraryJarFiles    = new ArrayList();
     private final List          libraryJarFilters  = new ArrayList();
-    private final List          configurationFiles = new ArrayList();
 
     // Accumulated configuration.
-    protected final Configuration configuration      = new Configuration();
+    protected final Configuration             configuration     = new Configuration();
+    private   final ProGuardTaskConfiguration taskConfiguration = ProGuardTaskConfiguration.create(getConfigurationProvider(), getObjectFactory());
 
     // Fields acting as parameters for the class member specification methods.
     private boolean            allowValues;
@@ -63,21 +68,25 @@ public abstract class ProGuardTask extends DefaultTask
     private static final String CONFIGURATION_FILE_NAME_PREFIX      = "/proguard/gradle/proguard-";
     private              String resolvedConfigurationFileNamePrefix = getProject().file(CONFIGURATION_FILE_NAME_PREFIX).toString();
 
+    @Inject
+    protected abstract ObjectFactory getObjectFactory();
+
+    @Nested
+    public ProGuardTaskConfiguration getTaskConfiguration()
+    {
+        return taskConfiguration;
+    }
 
     // Gradle task inputs and outputs, because annotations on the List fields
     // (private or not) don't seem to work. Private methods don't work either,
     // but package visible or protected methods are ok.
 
+    // Need to track injar and libraryjar files separately, even though they are also tracked as part of ProGuardTaskConfiguration.
+    // This is required to correctly add task dependencies for any file inputs, without requiring loading the configuration files early.
     @Classpath
     protected FileCollection getInJarFileCollection()
     {
         return getProject().files(inJarFiles);
-    }
-
-    @OutputFiles
-    protected FileCollection getOutJarFileCollection()
-    {
-        return getProject().files(outJarFiles);
     }
 
     @Classpath
@@ -86,12 +95,17 @@ public abstract class ProGuardTask extends DefaultTask
         return getProject().files(libraryJarFiles);
     }
 
-    @InputFiles
-    @PathSensitive(PathSensitivity.RELATIVE)
-    protected FileCollection getConfigurationFileCollection()
+    // Do not need to track outputs separately, since they do not carry task dependency information.
+    @Internal
+    protected FileCollection getOutJarFileCollection()
     {
-        return getProject().files(configurationFiles);
+        return getProject().files(outJarFiles);
     }
+
+    // Do not need to track configuration files directly, since they are mapped to the `Configuration` instance.
+    // The resulting `Configuration` instance is tracked as a task input.
+    @Internal
+    protected abstract ConfigurableFileCollection getConfigurationFiles();
 
     // Convenience methods to retrieve settings from outside the task.
 
@@ -109,7 +123,7 @@ public abstract class ProGuardTask extends DefaultTask
      * Returns the collected list of filters (represented as argument Maps)
      * corresponding to the list of input files.
      */
-    @Input
+    @Internal
     public List getInJarFilters()
     {
         return inJarFilters;
@@ -129,7 +143,7 @@ public abstract class ProGuardTask extends DefaultTask
      * Returns the collected list of filters (represented as argument Maps)
      * corresponding to the list of output files.
      */
-    @Input
+    @Internal
     public List getOutJarFilters()
     {
         return outJarFilters;
@@ -163,21 +177,12 @@ public abstract class ProGuardTask extends DefaultTask
      * Returns the collected list of filters (represented as argument Maps)
      * corresponding to the list of library files.
      */
-    @Input
+    @Internal
     public List getLibraryJarFilters()
     {
         return libraryJarFilters;
     }
 
-    /**
-     * Returns the collected list of configuration files to be included
-     * (represented as Object, String, File, etc).
-     */
-    @Internal
-    public List getConfigurationFiles()
-    {
-        return configurationFiles;
-    }
 
 
     // Gradle task settings corresponding to all ProGuard options.
@@ -185,16 +190,7 @@ public abstract class ProGuardTask extends DefaultTask
     public void configuration(Object configurationFiles)
     throws ParseException, IOException
     {
-        // Just collect the arguments, so they can be resolved lazily.
-        // Flatten collections, so they are easier to analyze later on.
-        if (configurationFiles instanceof Collection)
-        {
-            this.configurationFiles.addAll((Collection)configurationFiles);
-        }
-        else
-        {
-            this.configurationFiles.add(configurationFiles);
-        }
+        getConfigurationFiles().from(configurationFiles);
     }
 
     public void injars(Object inJarFiles)
@@ -1384,17 +1380,44 @@ public abstract class ProGuardTask extends DefaultTask
         loggingManager.captureStandardError(LogLevel.WARN);
 
         // Run ProGuard with the collected configuration.
-        new ProGuard(getConfiguration()).execute();
+        new ProGuard(getConfigurationProvider().get()).execute();
 
     }
 
+    private Provider<Configuration> getConfigurationProvider()
+    {
+        return getConfigurationFiles().getElements().map(this::getConfiguration);
+    }
+
+    private synchronized Configuration getConfiguration(Collection<FileSystemLocation> configurationFiles)
+    {
+        if (configuration.programJars == null)
+        {
+            try
+            {
+                loadConfiguration(configurationFiles);
+            }
+            catch (IOException | ParseException e)
+            {
+                throw new RuntimeException(e);
+            }
+        }
+        return configuration;
+    }
 
     /**
      * Returns the configuration collected so far, resolving files and
      * reading included configurations.
      */
-    private Configuration getConfiguration() throws IOException, ParseException
+    private Configuration loadConfiguration(Collection<FileSystemLocation> configurationFiles)
+    throws IOException, ParseException
     {
+        if (configuration.programJars != null)
+        {
+            return configuration;
+        }
+        getLogger().debug("Loading ProGuard configuration. This should happen immediately prior to ProGuardTask execution.");
+
         // Weave the input jars and the output jars into a single class path,
         // with lazy resolution of the files.
         configuration.programJars = new ClassPath();
@@ -1441,68 +1464,58 @@ public abstract class ProGuardTask extends DefaultTask
         }
 
         // Lazily apply the external configuration files.
-        for (int index = 0; index < configurationFiles.size(); index++)
+        for (FileSystemLocation configFile : configurationFiles)
         {
-            Object fileObject = configurationFiles.get(index);
 
-            ConfigurableFileCollection fileCollection =
-                getProject().files(fileObject);
+            File file = configFile.getAsFile();
 
-            // Parse the configuration as a collection of files.
-            Iterator<File> files = fileCollection.iterator();
-
-            while (files.hasNext())
+            // Check if this is the name of an internal configuration file.
+            if (isInternalConfigurationFile(file))
             {
-                File file = files.next();
+                getLogger().info("Loading default configuration file " +
+                                 file.getAbsolutePath());
 
-                // Check if this is the name of an internal configuration file.
-                if (isInternalConfigurationFile(file))
+                String internalConfigFileName =
+                    internalConfigurationFileName(file);
+
+                URL configurationURL =
+                    ProGuardTask.class.getResource(internalConfigFileName);
+
+                if (configurationURL == null)
                 {
-                    getLogger().info("Loading default configuration file " +
-                                     file.getAbsolutePath());
-
-                    String internalConfigFileName =
-                        internalConfigurationFileName(file);
-
-                    URL configurationURL =
-                        ProGuardTask.class.getResource(internalConfigFileName);
-
-                    if (configurationURL == null)
-                    {
-                        throw new FileNotFoundException("'" + file.getAbsolutePath() + "'");
-                    }
-
-                    // Parse the configuration as a URL.
-                    ConfigurationParser parser =
-                        new ConfigurationParser(configurationURL,
-                                                System.getProperties());
-
-                    try
-                    {
-                        parser.parse(configuration);
-                    }
-                    finally
-                    {
-                        parser.close();
-                    }
+                    throw new FileNotFoundException("'" + file.getAbsolutePath() + "'");
                 }
-                else
+
+                // Parse the configuration as a URL.
+                ConfigurationParser parser =
+                    new ConfigurationParser(configurationURL,
+                                            System.getProperties());
+
+                try
                 {
-                    getLogger().info("Loading configuration file " +
-                                     file.getAbsolutePath());
+                    parser.parse(configuration);
+                }
+                finally
+                {
+                    parser.close();
+                }
+            }
+            else
+            {
+                getLogger().info("Loading configuration file " +
+                        file.getAbsolutePath());
 
-                    ConfigurationParser parser =
+                ConfigurationParser parser =
                         new ConfigurationParser(file,
-                                                System.getProperties());
+                                System.getProperties());
 
-                    try
-                    {
-                        parser.parse(configuration);
-                    }
-                    finally
-                    {
-                        parser.close();
-                    }
+                try
+                {
+                    parser.parse(configuration);
+                }
+                finally
+                {
+                    parser.close();
                 }
             }
         }